# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Happy House API Documentation",
        "description": "",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/user/login": {
            "post": {
                "tags": [
                    "[Chủ nhà][Authentication]"
                ],
                "summary": "User login to system",
                "description": "Returns login data",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "[Chủ nhà][Authentication]"
                ],
                "summary": "User register to system",
                "description": "Returns register data",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user/forgot-password": {
            "post": {
                "tags": [
                    "[Chủ nhà][Authentication]"
                ],
                "summary": "User forgot password",
                "description": "Returns status forgot password",
                "operationId": "forgot-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user/reset-password": {
            "post": {
                "tags": [
                    "[Chủ nhà][Authentication]"
                ],
                "summary": "User reset password",
                "description": "Returns status reset password",
                "operationId": "reset-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user/change-password": {
            "put": {
                "tags": [
                    "[Chủ nhà][Authentication]"
                ],
                "summary": "User change password",
                "description": "Returns status change password",
                "operationId": "change-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/profile": {
            "get": {
                "tags": [
                    "[Chủ nhà][Authentication]"
                ],
                "summary": "User get profile",
                "description": "Returns user profile",
                "operationId": "get-profile",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/upload": {
            "post": {
                "tags": [
                    "[Upload] Upload file"
                ],
                "summary": "Upload single file",
                "description": "Upload single file",
                "operationId": "uploadFile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserChangePasswordRequest": {
                "title": "Change password request",
                "description": "Change password body data",
                "required": [
                    "current_password",
                    "new_password"
                ],
                "properties": {
                    "current_password": {
                        "type": "string",
                        "example": "password"
                    },
                    "new_password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserForgotPasswordRequest": {
                "title": "Forgot password request",
                "description": "Forgot password body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "user@gmail.com"
                    }
                },
                "type": "object"
            },
            "UserLoginRequest": {
                "title": "Login request",
                "description": "Login body data",
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "title": "username",
                        "description": "Username or phone of user",
                        "type": "string",
                        "example": "user@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserRegisterRequest": {
                "title": "Register request",
                "description": "Register body data",
                "required": [
                    "name",
                    "phone",
                    "email",
                    "password",
                    "location",
                    "register_type"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of user",
                        "type": "string",
                        "example": "Nguyen van A"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "string",
                        "example": "09498487494"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "xxxx@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "dfngker983"
                    },
                    "address": {
                        "title": "address",
                        "description": "address of user",
                        "type": "string",
                        "example": "xóm 4 hải hâu"
                    },
                    "invite_code": {
                        "title": "invite_code",
                        "description": "invite_code of user",
                        "type": "string",
                        "example": "nmdf9g34nf"
                    }
                },
                "type": "object"
            },
            "UserResetPasswordRequest": {
                "title": "Reset password request",
                "description": "Reset password body data",
                "required": [
                    "email",
                    "token",
                    "password",
                    "password_confirmed"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "example": "user@gmail.com"
                    },
                    "token": {
                        "title": "token",
                        "description": "token reset",
                        "example": "efgjej3$deeddd"
                    },
                    "password": {
                        "title": "password",
                        "description": "new password",
                        "example": "password"
                    },
                    "password_confirmed": {
                        "title": "password_confirmed",
                        "description": "password confirmed",
                        "example": "password"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token\\>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ],
    "security": [
        []
    ]
}